+++ scripts/common.js
@@ -1,26 +1,4 @@
-import {
-  sampleRUM,
-  loadCSS,
-  loadBlock,
-  loadBlocks,
-  loadHeader,
-  buildBlock,
-  decorateBlock,
-} from './lib-franklin.js';
-
 let placeholders = null;
-
-/**
-   * loads a block named 'footer' into footer
-   */
-function loadFooter(footer) {
-  if (footer) {
-    const footerBlock = buildBlock('footer', '');
-    footer.append(footerBlock);
-    decorateBlock(footerBlock);
-    loadBlock(footerBlock);
-  }
-}
 
 export async function getPlaceholders() {
   const language = window.location.pathname.match(/\/fr\//);
@@ -76,86 +54,6 @@ export function addFavIcon(href) {
   }
 }
 
-export async function loadTemplate(doc, templateName) {
-  try {
-    const cssLoaded = new Promise((resolve) => {
-      loadCSS(`${window.hlx.codeBasePath}/templates/${templateName}/${templateName}.css`, resolve);
-    });
-    const decorationComplete = new Promise((resolve) => {
-      (async () => {
-        try {
-          const mod = await import(`../templates/${templateName}/${templateName}.js`);
-          if (mod.default) {
-            await mod.default(doc);
-          }
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.log(`failed to load module for ${templateName}`, error);
-        }
-        resolve();
-      })();
-    });
-    await Promise.all([cssLoaded, decorationComplete]);
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.log(`failed to load block ${templateName}`, error);
-  }
-}
-
-/**
-   * loads everything that doesn't need to be delayed.
-   */
-export async function loadLazy(doc) {
-  const main = doc.querySelector('main');
-  await loadBlocks(main);
-
-  const { hash } = window.location;
-  const element = hash ? doc.getElementById(hash.substring(1)) : false;
-  if (hash && element) element.scrollIntoView();
-  const header = doc.querySelector('header');
-
-  loadHeader(header);
-  loadFooter(doc.querySelector('footer'));
-
-  const subnav = header?.querySelector('.block.sub-nav');
-  if (subnav) {
-    loadBlock(subnav);
-  }
-
-  loadCSS(`${window.hlx.codeBasePath}/styles/lazy-styles.css`);
-  addFavIcon(`${window.hlx.codeBasePath}/styles/favicon.svg`);
-  sampleRUM('lazy');
-  sampleRUM.observe(main.querySelectorAll('div[data-block-name]'));
-  sampleRUM.observe(main.querySelectorAll('picture > img'));
-}
-
-/**
-   * loads everything that happens a lot later, without impacting
-   * the user experience.
-   */
-export function loadDelayed() {
-  // eslint-disable-next-line import/no-cycle
-  window.setTimeout(() => {
-    import('./delayed.js');
-  }, 3000);
-  // load anything that can be postponed to the latest here
-}
-
-export const removeEmptyTags = (block) => {
-  block.querySelectorAll('*').forEach((x) => {
-    const tagName = `</${x.tagName}>`;
-
-    // checking that the tag is not autoclosed to make sure we don't remove <meta />
-    // checking the innerHTML and trim it to make sure the content inside the tag is 0
-    if (
-      x.outerHTML.slice(tagName.length * -1).toUpperCase() === tagName
-        // && x.childElementCount === 0
-        && x.innerHTML.trim().length === 0) {
-      x.remove();
-    }
-  });
-};
-
 export const unwrapDivs = (element) => {
   Array.from(element.children).forEach((node) => {
     if (node.tagName === 'DIV' && node.attributes.length === 0) {
@@ -170,75 +68,6 @@ export const unwrapDivs = (element) => {
   });
 };
 
-export const variantsClassesToBEM = (blockClasses, expectedVariantsNames, blockName) => {
-  expectedVariantsNames.forEach((variant) => {
-    if (blockClasses.contains(variant)) {
-      blockClasses.remove(variant);
-      blockClasses.add(`${blockName}--${variant}`);
-    }
-  });
-};
-
-/**
-   *
-   * @param {string} blockName - block name with '-' instead of spaces
-   * @param {string} blockContent - the content that will be set as block inner HTML
-   * @param {object} options - other options like variantsClasses
-   * @returns
-   */
-export async function loadAsBlock(blockName, blockContent, options = {}) {
-  const { variantsClasses = [] } = options;
-  const blockEl = createElement('div', {
-    classes: ['block', blockName, ...variantsClasses],
-    props: { 'data-block-name': blockName },
-  });
-  const wrapperEl = createElement('div');
-  wrapperEl.append(blockEl);
-
-  blockEl.innerHTML = blockContent;
-  await loadBlocks(wrapperEl);
-
-  return blockEl;
-}
-
-export const adjustPretitle = (element) => {
-  const headingSelector = 'h1, h2, h3, h4, h5, h6';
-
-  [...element.querySelectorAll(headingSelector)].forEach((heading) => {
-    const isNextElHeading = heading.nextElementSibling?.matches(headingSelector);
-
-    if (!isNextElHeading) {
-      return;
-    }
-
-    const currentLevel = Number(heading.tagName[1]);
-    const nextElLevel = Number(heading.nextElementSibling.tagName[1]);
-
-    if (currentLevel > nextElLevel) {
-      const pretitle = createElement('span', { classes: ['pretitle'] });
-      pretitle.append(...heading.childNodes);
-
-      heading.replaceWith(pretitle);
-    }
-  });
-};
-
-export const slugify = (text) => (
-  text.toString().toLowerCase().trim()
-  // separate accent from letter
-    .normalize('NFD')
-  // remove all separated accents
-    .replace(/[\u0300-\u036f]/g, '')
-  // replace spaces with -
-    .replace(/\s+/g, '-')
-  // replace & with 'and'
-    .replace(/&/g, '-and-')
-  // remove all non-word chars
-    .replace(/[^\w-]+/g, '')
-  // replace multiple '-' with single '-'
-    .replace(/--+/g, '-')
-);
-
 /**
    * Check if one trust group is checked.
    * @param {String} groupName the one trust croup like: C0002
@@ -247,50 +76,3 @@ export function checkOneTruckGroup(groupName) {
   const oneTrustCookie = decodeURIComponent(document.cookie.split(';').find((cookie) => cookie.trim().startsWith('OptanonConsent=')));
   return oneTrustCookie.includes(`${groupName}:1`);
 }
-
-export function isEloquaFormAllowed() {
-  return checkOneTruckGroup('C0004');
-}
-
-/*
-    The generateId function should be used only
-    for generating the id for UI elements
-  */
-let idValue = 0;
-
-export const generateId = (prefix = 'id') => {
-  idValue += 1;
-  return `${prefix}-${idValue}`;
-};
-
-/**
-   * Helper for delaying a function
-   * @param {function} func callback function
-   * @param {number} timeout time to debouce in ms, default 200
-  */
-export function debounce(func, timeout = 200) {
-  let timer;
-  return (...args) => {
-    clearTimeout(timer);
-
-    timer = setTimeout(() => { func.apply(this, args); }, timeout);
-  };
-}
-
-/**
-   * Returns a list of properties listed in the block
-   * @param {string} route get the Json data from the route
-   * @returns {Object} the json data object
-  */
-export const getJsonFromUrl = async (route) => {
-  try {
-    const response = await fetch(route);
-    if (!response.ok) return null;
-    const json = await response.json();
-    return json;
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('getJsonFromUrl:', { error });
-  }
-  return null;
-};
