--- styles/lazy-styles.css
@@ -1,51 +1,0 @@
-/* below the fold CSS goes here */
-
-/* fonts based on https://fonts.google.com/specimen/Roboto, only including latin versions to reduce size */
-@font-face {
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: 300;
-    font-display: swap;
-    src: url("../fonts/Roboto-300.woff2") format('woff2');
-    unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA, U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215, U+FEFF, U+FFFD;
-}
-
-@font-face {
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: 500;
-    font-display: swap;
-    src: url("../fonts/Roboto-500.woff2") format('woff2');
-    unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA, U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215, U+FEFF, U+FFFD;
-}
-
-@font-face {
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: 900;
-    font-display: swap;
-    src: url("../fonts/Roboto-900.woff2") format('woff2');
-    unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA, U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215, U+FEFF, U+FFFD;
-}
-
-@font-face {
-    font-family: 'Roboto Condensed';
-    font-style: normal;
-    font-weight: 300;
-    font-display: swap;
-    src: url("../fonts/Roboto-Condensed-300.woff2") format('woff2');
-    unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA, U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215, U+FEFF, U+FFFD;
-}
-
-.section#whypartsasist {
-    background-image: url("/media/parallax1.png"), url("/media/parallax2.png");
-    background-repeat: no-repeat, no-repeat;
-    background-position: top left, right;
-}
-
-main .section:last-child {
-    /* reduce padding of the last section in main. Instead we increase the padding-top in the footer
-     and show more of background image. */
-    padding-bottom: 0;
-}
-
