+++ styles/styles.css
@@ -84,7 +84,7 @@ body {
 }
 
 body.appear {
-  display: unset;
+  display: block;
 }
 
 h1, h2, h3,
