+++ test/scripts/scripts.test.js
@@ -16,7 +16,7 @@ document.head.innerHTML = await readFile({ path: './head.html' });
 describe('Core Helix features', () => {
   before(async () => {
     scripts = await import('../../scripts/scripts.js');
-    lib = await import('../../scripts/lib-franklin.js');
+    lib = await import('../../scripts/aem.js');
 
     document.body.innerHTML = await readFile({ path: './body.html' });
   });
